containers/terraform-triggers/tests/requirements.txt
boto3==1.28.67
nats-py==2.6.0
